Add files via upload
Data clean-up and pre-processing is carried out. The pre-processed data is split into testing and training sets. Keras sequential model is built to train the data, after which testing phase is carried out. The inputs to the model are {passenger class, age, fare} and the predict output={sex - male of female}. Later the accuracy of the models is checked by creating the confusion matrix, and classification report (precision and recall).